+++ GRDB/Core/DatabaseRegion.swift
@@ -55,14 +55,6 @@ public struct DatabaseRegion: CustomStringConvertible, Equatable {
     /// from all tables.
     public static let fullDatabase = DatabaseRegion(tableRegions: nil)
     
-    // TODO: rename to init(table:) when this initializer is no longer public
-    /// The region that covers a full database table: all columns and all rows
-    /// from this table.
-    static func fullTable(_ table: String) -> DatabaseRegion {
-        let table = CaseInsensitiveIdentifier(rawValue: table)
-        return DatabaseRegion(tableRegions: [table: TableRegion(columns: nil, rowIds: nil)])
-    }
-    
     /// Creates an empty database region.
     public init() {
         self.init(tableRegions: [:])
@@ -71,9 +63,9 @@ public struct DatabaseRegion: CustomStringConvertible, Equatable {
     /// Creates a region that spans all rows and columns of a database table.
     ///
     /// - parameter table: A table name.
-    @available(*, deprecated, message: "In order to specify a table region, prefer `Table(tableName)`")
-    public init(table: String) {
-        self = .fullTable(table)
+    init(table: String) {
+        let table = CaseInsensitiveIdentifier(rawValue: table)
+        self.init(tableRegions: [table: TableRegion(columns: nil, rowIds: nil)])
     }
     
     /// Full columns in a table: (some columns in a table) × (all rows)