+++ GRDB/FTS/FTS3.swift
@@ -79,14 +79,6 @@ public struct FTS3: VirtualTableModule {
     /// The virtual table module name
     public let moduleName = "fts3"
     
-    // TODO: remove when `makeTableDefinition()` is no longer a requirement
-    /// Reserved; part of the VirtualTableModule protocol.
-    ///
-    /// See Database.create(virtualTable:using:)
-    public func makeTableDefinition() -> FTS3TableDefinition {
-        preconditionFailure()
-    }
-    
     /// Reserved; part of the VirtualTableModule protocol.
     ///
     /// See Database.create(virtualTable:using:)