+++ GRDB/QueryInterface/SQL/SQLFunctions.swift
@@ -26,17 +26,6 @@ public func average(_ value: SQLSpecificExpressible) -> SQLExpression {
 ///
 ///     // COUNT(email)
 ///     count(Column("email"))
-///
-/// :nodoc:
-@available(*, deprecated)
-public func count(_ counted: SQLSelectable) -> SQLExpression {
-    counted.sqlSelection.countExpression
-}
-
-/// Returns an expression that evaluates the `COUNT` SQL function.
-///
-///     // COUNT(email)
-///     count(Column("email"))
 public func count(_ counted: SQLSpecificExpressible) -> SQLExpression {
     .count(counted.sqlExpression)
 }