+++ CHANGELOG.md
@@ -97,6 +97,7 @@ GRDB adheres to [Semantic Versioning](https://semver.org/), with one exception: 
 
 - **Breaking**: Swift 5.7+ and Xcode 14+ are required.
 - **Breaking**: iOS 11.0+ / macOS 10.13+ / tvOS 11.0+ / watchOS 4.0+ are required.
+- **Breaking**: Removed deprecated methods
 
 ## 5.26.0